+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [10.13.0] - 26-03-27
+
+### Fixed
+
+- Fixed an issue on Android where the Ads API would sometimes return an empty map instead of an array when querying the list of current ads or scheduled ad breaks.
+- Fixed an issue on iOS where the transition from PiP to fullscreen was not smooth.
+
 ## [10.12.1] - 26-03-19
 
 ### Fixed